Exception handling

+++ src/main/java/br/com/rgsystems/dscomerce/services/ProductService.java
@@ -3,10 +3,15 @@ package br.com.rgsystems.dscomerce.services;
 import br.com.rgsystems.dscomerce.dto.ProductDTO;
 import br.com.rgsystems.dscomerce.entities.Product;
 import br.com.rgsystems.dscomerce.repositories.ProductRepository;
+import br.com.rgsystems.dscomerce.services.exceptions.DatabaseException;
+import br.com.rgsystems.dscomerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -21,7 +26,7 @@ public class ProductService {
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
         Optional<Product> result = repository.findById(id);
-        Product product = result.get();
+        Product product = result.orElseThrow(() -> new ResourceNotFoundException("Cadastro não localizado"));
         ProductDTO dto = new ProductDTO(product);
         return dto;
     }
@@ -41,15 +46,29 @@ public class ProductService {
 
     @Transactional
     public ProductDTO  update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDTOToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDTOToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+
     }
 
-    @Transactional()
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDTOToEntity(ProductDTO dto, Product entity) {
